# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Education Mini Program API
  version: 0.1.0
  description: >-
    Shared API contract for the first-phase education mini program. Successful
    responses use a data envelope and failures preserve their HTTP status.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Environment-relative service root
...
security:
  - bearerAuth: []
paths:
...
  /campus-managers/me/dashboard:
    get:
      operationId: getCampusManagerDashboard
      summary: Read the current campus dashboard
      tags: [CampusManager]
      responses:
        '200':
          description: Campus dashboard data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerDashboardEnvelope'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/profile:
    get:
      operationId: getCampusManagerProfile
      summary: Read the current campus manager profile
      tags: [CampusManager]
      responses:
        '200':
          description: Campus manager profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerProfileEnvelope'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/students:
    get:
      operationId: listCampusManagerStudents
      summary: List active students in the current campus
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: query
          in: query
          required: false
          schema: { type: string, maxLength: 100 }
      responses:
        '200':
          description: Campus student page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerStudentsPageEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createCampusManagerStudent
      summary: Create a student in the current campus
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampusManagerCreateStudentRequest'
      responses:
        '200':
          description: Student created or replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerStudentEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/students/{studentId}:
    get:
      operationId: getCampusManagerStudent
      summary: Read a student in the current campus
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '200':
          description: Campus student detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerStudentDetailEnvelope'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/scheduling-options:
    get:
      operationId: getCampusManagerSchedulingOptions
      summary: Read active classes and derived teachers for scheduling
      tags: [CampusManager]
      responses:
        '200':
          description: Current-campus scheduling options.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerSchedulingOptionsEnvelope'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/lesson-sessions:
    get:
      operationId: listCampusManagerLessonSessions
      summary: List lesson sessions in the current campus
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/FromDateTime'
        - $ref: '#/components/parameters/ToDateTime'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/LessonSessionStatus'
      responses:
        '200':
          description: Campus lesson page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerLessonSessionsPageEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createCampusManagerLessonSession
      summary: Create one future lesson session in the current campus
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampusManagerCreateLessonSessionRequest'
      responses:
        '200':
          description: Lesson session created or replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerLessonSessionEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/lesson-sessions/{lessonSessionId}:
    patch:
      operationId: updateCampusManagerLessonSession
      summary: Update one scheduled lesson session
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/LessonSessionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampusManagerVersionedUpdateLessonSessionRequest'
      responses:
        '200':
          description: Lesson session updated or replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerLessonSessionEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/lesson-sessions/{lessonSessionId}/cancel:
    post:
      operationId: cancelCampusManagerLessonSession
      summary: Cancel one scheduled lesson session
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/LessonSessionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampusManagerVersionedCancelLessonSessionRequest'
      responses:
        '200':
          description: Lesson session cancelled or replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerLessonSessionEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/leave-requests:
    get:
      operationId: listCampusManagerLeaveRequests
      summary: List current-campus leave requests
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ParentLeaveStatus'
      responses:
        '200':
          description: Campus leave request page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerLeaveRequestsPageEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/leave-requests/{leaveRequestId}/approve:
    post:
      operationId: approveCampusManagerLeaveRequest
      summary: Approve one pending current-campus leave request
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: leaveRequestId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampusManagerVersionedReviewLeaveRequest'
      responses:
        '200':
          description: Leave request approved or replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerLeaveRequestEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/leave-requests/{leaveRequestId}/reject:
    post:
      operationId: rejectCampusManagerLeaveRequest
      summary: Reject one pending current-campus leave request
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: leaveRequestId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampusManagerVersionedRejectLeaveRequest'
      responses:
        '200':
          description: Leave request rejected or replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerLeaveRequestEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/warnings:
    get:
      operationId: listCampusManagerWarnings
      summary: List low-balance students in the current campus
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: query
          in: query
          required: false
          schema: { type: string, maxLength: 100 }
      responses:
        '200':
          description: Current-campus warning page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerWarningsPageEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /campus-managers/me/campus:
    get:
      operationId: getCampusManagerCampus
      summary: Read editable settings for the current campus
      tags: [CampusManager]
      responses:
        '200':
          description: Current campus settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerCampusEnvelope'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    patch:
      operationId: updateCampusManagerCampus
      summary: Update editable settings for the current campus
      tags: [CampusManager]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampusManagerVersionedUpdateCampusRequest'
      responses:
        '200':
          description: Current campus settings updated or replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampusManagerCampusEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by the authentication module.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Caller-generated key used by mutation endpoints to make retries safe.
        Keys contain 8-128 letters, numbers, dots, underscores, colons, or
        hyphens.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]{8,128}$'
      example: lesson-confirm-20260827-0001
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    FromDateTime:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: Inclusive range start in RFC 3339 format with an explicit offset.
    ToDateTime:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: Exclusive range end in RFC 3339 format with an explicit offset.
    LessonSessionId:
      name: lessonSessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    StudentId:
      name: studentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    RequestId:
      description: Correlation identifier matching the response body requestId.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
...
  responses:
...
    BadRequest:
      description: Request validation failed.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Authentication is required or the session is invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: The authenticated user lacks the required role or data scope.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: The requested state transition conflicts with current state.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ResourceNotFound:
      description: The requested resource does not exist.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            missingResource:
              value:
                code: RESOURCE_NOT_FOUND
                message: Cannot GET /resource
                requestId: 64ea9657-41f1-4baf-a3fb-d62e39054016
                details: {}
    InternalServerError:
      description: An unexpected server error occurred.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            internalError:
              value:
                code: INTERNAL_SERVER_ERROR
                message: Internal server error
                requestId: 64ea9657-41f1-4baf-a3fb-d62e39054016
                details: {}
  schemas:
...
    SuccessEnvelope:
      type: object
      description: Common wrapper for every successful response.
      required:
        - data
        - requestId
      properties:
        data:
          description: Endpoint-specific response payload.
        requestId:
          type: string
          minLength: 1
          maxLength: 128
          description: Correlation identifier returned in X-Request-Id.
      additionalProperties: false
    PaginatedSuccessEnvelope:
      type: object
      description: Common wrapper for paginated list responses.
      required:
        - data
        - meta
        - requestId
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        requestId:
          type: string
          minLength: 1
          maxLength: 128
          description: Correlation identifier returned in X-Request-Id.
      additionalProperties: false
...
    ErrorEnvelope:
      type: object
      description: Common response body for failures; the HTTP status is preserved.
      required:
        - code
        - message
        - requestId
        - details
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
          maxLength: 128
        details:
          type: object
          additionalProperties: true
      additionalProperties: false
...
    LessonSessionStatus:
      type: string
      enum:
        - SCHEDULED
        - IN_PROGRESS
        - COMPLETED
        - REVERSED
        - CANCELLED
...
    LessonKind:
      type: string
      enum:
        - REGULAR
        - MAKEUP
        - TRIAL
...
    ParentLeaveStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
...
    CampusManagerDashboardView:
      type: object
      required:
        - manager
        - campus
        - activeStudentCount
        - todayLessonCount
        - pendingLeaveCount
        - latestPendingLeave
        - serverTime
      properties:
        manager:
          type: object
          required: [displayName]
          properties:
            displayName: { type: string, minLength: 1, maxLength: 100 }
          additionalProperties: false
        campus:
          $ref: '#/components/schemas/CampusManagerCampusSummary'
        activeStudentCount: { type: integer, minimum: 0 }
        todayLessonCount: { type: integer, minimum: 0 }
        pendingLeaveCount: { type: integer, minimum: 0 }
        latestPendingLeave:
          oneOf:
            - $ref: '#/components/schemas/CampusManagerLatestLeaveView'
            - type: 'null'
        serverTime: { type: string, format: date-time }
      additionalProperties: false
    CampusManagerProfileView:
      type: object
      required: [displayName, roleCode, campusId, campusName]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 100 }
        roleCode:
          type: string
          enum: [CAMPUS_MANAGER]
        campusId: { type: string, format: uuid }
        campusName: { type: string, minLength: 1, maxLength: 200 }
      additionalProperties: false
    CampusManagerStudentSummary:
      type: object
      required: [id, campusId, displayName, birthDate, classNames]
      properties:
        id: { type: string, format: uuid }
        campusId: { type: string, format: uuid }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        birthDate:
          oneOf:
            - type: string
              format: date
            - type: 'null'
        classNames:
          type: array
          items: { type: string, minLength: 1, maxLength: 200 }
      additionalProperties: false
...
    CampusManagerCreateStudentRequest:
      type: object
      required: [displayName]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 100 }
        birthDate:
          oneOf:
            - type: string
              format: date
            - type: 'null'
        classGroupId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
      additionalProperties: false
...
    CampusManagerCreateLessonSessionRequest:
      type: object
      required: [classGroupId, startsAt, endsAt, kind]
      properties:
        classGroupId: { type: string, format: uuid }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        kind: { $ref: '#/components/schemas/LessonKind' }
      additionalProperties: false
    CampusManagerVersionedUpdateLessonSessionRequest:
      type: object
      required: [version, startsAt, endsAt, kind]
      properties:
        version: { type: integer, minimum: 1 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        kind: { $ref: '#/components/schemas/LessonKind' }
      additionalProperties: false
    CampusManagerVersionedCancelLessonSessionRequest:
      type: object
      required: [version]
      properties:
        version: { type: integer, minimum: 1 }
      additionalProperties: false
...
    CampusManagerVersionedReviewLeaveRequest:
      type: object
      required: [version]
      properties:
        version: { type: integer, minimum: 1 }
        reviewReason: { type: string, maxLength: 500 }
      additionalProperties: false
    CampusManagerVersionedRejectLeaveRequest:
      type: object
      required: [version, reviewReason]
      properties:
        version: { type: integer, minimum: 1 }
        reviewReason: { type: string, minLength: 1, maxLength: 500 }
      additionalProperties: false
...
    CampusManagerVersionedUpdateCampusRequest:
      type: object
      required: [version, name, contactPhone, address, lessonWarningThresholdUnits]
      properties:
        version: { type: integer, minimum: 1 }
        name: { type: string, minLength: 1, maxLength: 200, pattern: '.*\S.*' }
        contactPhone:
          oneOf:
            - type: string
              maxLength: 30
              pattern: '^(?:1[3-9][0-9]{9}|0[0-9]{2,3}-?[0-9]{7,8})$'
            - type: 'null'
        address:
          oneOf:
            - type: string
              maxLength: 300
            - type: 'null'
        lessonWarningThresholdUnits: { type: integer, minimum: 0 }
      additionalProperties: false
...
    CampusManagerDashboardEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CampusManagerDashboardView' }
    CampusManagerProfileEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CampusManagerProfileView' }
    CampusManagerStudentEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CampusManagerStudentSummary' }
    CampusManagerStudentDetailEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CampusManagerStudentDetail' }
    CampusManagerStudentsPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/PaginatedSuccessEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/CampusManagerStudentSummary' }
    CampusManagerSchedulingOptionsEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CampusManagerSchedulingOptionsView' }
    CampusManagerLessonSessionEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CampusManagerLessonSessionView' }
    CampusManagerLessonSessionsPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/PaginatedSuccessEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/CampusManagerLessonSessionView' }
    CampusManagerLeaveRequestEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CampusManagerLeaveRequestView' }
    CampusManagerLeaveRequestsPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/PaginatedSuccessEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/CampusManagerLeaveRequestView' }
    CampusManagerWarningsPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/PaginatedSuccessEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/CampusManagerWarningView' }
    CampusManagerCampusEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CampusManagerCampusView' }
...